Rename magic argument to no argument
