+++ lib/rubocop/cop/magic_numbers/no_argument.rb
