+++ lib/rubocop/no_magic_numbers.rb
@@ -2,3 +2,4 @@
 
 require 'rubocop/cop/magic_numbers/version'
 require 'rubocop/cop/magic_numbers/no_magic_numbers'
+require 'rubocop/cop/magic_numbers/no_argument'
